requirements.txt
Flask==2.3.3
supabase==2.0.2
python-dotenv==1.0.0
sendgrid==6.12.5
Pillow==10.0.1